Add when and always helpers to replace convergeOn export
